Change miles_from_campus to integer

+++ database/migrations/2023_07_17_013743_create_listings_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
 
             $table->integer('price');
             $table->boolean('available')->default(true);
-            $table->tinyInteger('miles_from_campus');
+            $table->integer('miles_from_campus');
 
             $table->foreignIdFor(User::class)->constrained()->cascadeOnDelete();
